Move a type to the integrations namespace
