+++ test/Altinn.Notifications.Tests/Notifications.Integrations/Register/RegisterClientTests.cs
@@ -11,6 +11,7 @@ using Altinn.Notifications.Core.Models.Parties;
 using Altinn.Notifications.Core.Shared;
 using Altinn.Notifications.Integrations.Configuration;
 using Altinn.Notifications.Integrations.Register;
+using Altinn.Notifications.Integrations.Register.Models;
 
 using Microsoft.Extensions.Options;
